functions/requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.14.0
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0
requests>=2.28.0
praw>=7.6.0
